+++ src/main/java/morgan/lesbois/interfaces/GrappleInterface.java
@@ -6,6 +6,6 @@ import org.jetbrains.annotations.Nullable;
 public interface GrappleInterface {
     public GrappleHookEntity lesbois$getGrappleHook();
     public void lesbois$setGrappleHook(@Nullable GrappleHookEntity hook);
-    public GrappleHookEntity lesbois$grapple(double maxDistance, double minDistance, boolean disableFallDamage, double pullSpeed, double lookAssist, double damping);
+    public GrappleHookEntity lesbois$grapple(float maxDistance, float minDistance, boolean disableFallDamage, float pullSpeed, float lookAssist, float damping);
     public boolean lesbois$unGrapple();
 }
